requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
